chore(add-criteria): isolate advanced search feature into its own component
