+++ src/searchModal/tpl/advanced-search.tpl
@@ -1,0 +1,9 @@
+<div class="advanced-search-container">
+    <div class="advanced-criteria-container"></div>
+    <div class="add-criteria-container">
+        <a><span class="icon-add"></span> add criteria</a>
+        <select name="criteria-select">
+            <option></option>
+        </select>
+    </div>
+</div>
