+++ src/searchModal/tpl/layout.tpl
@@ -14,16 +14,6 @@
                         <div class="class-tree"></div>
                     </div>
                 </div>
-                <!-- TODO: This container must only be rendered if advanced search is enabled -->
-                <div class="advanced-search-container">
-                    <div class="advanced-criteria-container"></div>
-                    <div class="add-criteria-container">
-                        <a><span class="icon-add"></span> add criteria</a>
-                        <select name="criteria-select">
-                            <option></option>
-                        </select>
-                    </div>
-                </div>
             </div>
             <div class="buttons-container">
                 <button class="btn-clear btn-transparent small">{{__ "Clear"}}</button>
